--- kafka-service/src/main/java/org/uniovi/asw/inci_manager/kafka/App.java
@@ -1,11 +1,0 @@
-package org.uniovi.asw.inci_manager.kafka;
-
-/**
- * Hello world!
- *
- */
-public class App {
-    public static void main(String[] args) {
-	System.out.println("Hello World!");
-    }
-}
