--- kafka-service/src/test/java/org/uniovi/asw/inci_manager/kafka/AppTest.java
@@ -1,38 +1,0 @@
-package org.uniovi.asw.inci_manager.kafka;
-
-import org.junit.experimental.categories.Category;
-
-import TestKit.UnitTest;
-import junit.framework.Test;
-import junit.framework.TestCase;
-import junit.framework.TestSuite;
-
-/**
- * Unit test for simple App.
- */
-@Category(UnitTest.class)
-public class AppTest extends TestCase {
-    /**
-     * Create the test case
-     *
-     * @param testName
-     *            name of the test case
-     */
-    public AppTest(String testName) {
-	super(testName);
-    }
-
-    /**
-     * @return the suite of tests being tested
-     */
-    public static Test suite() {
-	return new TestSuite(AppTest.class);
-    }
-
-    /**
-     * Rigourous Test :-)
-     */
-    public void testApp() {
-	assertTrue(true);
-    }
-}
