--- resources/application-test.properties
@@ -1,1 +1,0 @@
-agents.url=http://localhost:8080/user
