--- resources/application.properties
@@ -1,2 +1,0 @@
-server.port=8090
-agents.url=https://agents-module.herokuapp.com/user
