--- rest-service/src/main/java/org/uniovi/asw/inci_manager/rest/StartupREST.java
@@ -1,17 +1,0 @@
-package org.uniovi.asw.inci_manager.rest;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-/**
- * Hello world!
- *
- */
-@SpringBootApplication
-@ComponentScan(basePackages= {"org.uniovi.asw.inci_manager.rest"})
-public class StartupREST {
-	public static void main( String[] args ) {
-		SpringApplication.run(StartupREST.class, args);
-	}
-}
