--- rest-service/src/main/resources/application-test.properties
@@ -1,2 +1,0 @@
-server.port=8090
-agents.url=http://localhost:8080/user
