--- rest-service/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-server.port=8080
-agents.url=https://agents-module.herokuapp.com/user
