+++ src/main/java/org/uniovi/i3a/incimanager/Application.java
@@ -1,0 +1,15 @@
+package org.uniovi.i3a.incimanager;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * Hello world!
+ *
+ */
+@SpringBootApplication
+public class Application {
+    public static void main(String[] args) {
+	SpringApplication.run(Application.class, args);
+    }
+}
