--- web-service/src/main/resources/master.csv
@@ -1,3 +1,0 @@
-3,Sensor
-2,Entity
-1,Person
